+++ unittest/libmariadb/logs.c
@@ -61,7 +61,7 @@ static int test_logs(MYSQL *mysql)
   MYSQL_STMT *stmt;
   MYSQL_BIND my_bind[2];
   char       data[255];
-  ulong      length;
+  size_t     length;
   int        rc;
   short      id;
